Known variables cannot be parsed
Many known variables contain the word `is` which is also a keyword of the language. Unfortunately, using `is` as a keyword is too restrictive linguistically and prevents the use of natural language variable names. Replacing `is` with `=` or `equals` would imply that equality is a valid comparison when it is in fact not present in the language and I had found a tendency to even attempt to use `is` as an equality comparison.

As a result `:=` is chosen as the assignment keyword because it cannot be part of a variable name and isn't close enough to `=` to be confused as a comparison.